evm: fix deployment script

+++ evm/script/DeployWormholeNtt.s.sol
@@ -24,7 +24,7 @@ contract DeployWormholeNtt is Script, ParseNttConfig {
         address specialRelayerAddr;
         uint8 consistencyLevel;
         uint256 gasLimit;
-        uint64 outboundLimit;
+        uint256 outboundLimit;
     }
 
     // The minimum gas limit to verify a message on mainnet. If you're worried about saving
@@ -81,7 +81,7 @@ contract DeployWormholeNtt is Script, ParseNttConfig {
     function configureNttManager(
         address nttManager,
         address transceiver,
-        uint64 outboundLimit,
+        uint256 outboundLimit,
         bool shouldSkipRateLimiter
     ) public {
         IManagerBase(nttManager).setTransceiver(transceiver);
@@ -133,7 +133,7 @@ contract DeployWormholeNtt is Script, ParseNttConfig {
         require(params.gasLimit >= MIN_WORMHOLE_GAS_LIMIT, "Invalid gas limit");
 
         // Outbound rate limiter limit.
-        params.outboundLimit = uint64(vm.envUint("RELEASE_OUTBOUND_LIMIT"));
+        params.outboundLimit = vm.envUint("RELEASE_OUTBOUND_LIMIT");
     }
 
     function run() public {
